popUpForm working with dummy data, separate folder created
